assets/module4/front/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.6.2
    # via flask
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via flask
flask==2.3.2
    # via -r requirements.in
google-auth==2.19.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
requests==2.31.0
    # via -r requirements.in
rsa==4.9
    # via google-auth
six==1.16.0
    # via google-auth
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
werkzeug==2.3.4
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
